requirements.txt
datasets
tqdm
tabulate
huggingface-hub
typing-extensions
argparse
ollama
torch
pillow
streamlit
watchdog
graphviz
